first working version
Well, it parses cmd line args, needed files and talks to GH APIv3. It is also
capable of labeling issues that should be labeled according to label file.

It doesnt handle Link header which GH includes when pagination is in effect.
That means currently you will only be able to list first 30 issues for a given
repository. Next commit will be fixing this. I promise