fix: repo with username
I previously did the repo option only just for repo name without
realizing that the username will be hardcoded and thus it will never ever,
like never never ever, be able to work. My apologies. Now the repo option
also takes a username - in the usual GH URL format, that means it precedes
the repo name and is separated from it by a slash. Classics